requirements.txt
pandas>=2.0
openpyxl>=3.0
python-docx>=0.8
docx2txt>=0.8
